+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Finance: Create UBL invoices for Peppol
 - Addressbook: Fixed contact sorting on last name
 - E-mail: Fixed option "Sort on last contact time"
+- Tasks / support: group task list combo by task list grouping.
 
 24-06-2024: 6.8.56
 - E-mail: Fixed autocomplete bug where typed text would stick too